feat(nutrition-domain): extract pure helpers into @sergeant/nutrition-domain
Phase 7 / PR 1: перший крок порту модуля Харчування у RN — витяг
DOM-free помічників у окремий workspace-пакет за зразком
`@sergeant/{finyk,fizruk,routine}-domain`.

Перенесено (rename + оновлені імпорти з `@sergeant/shared`):
- `mealTypes` (+ test)
- `pantryTextParser` (+ test)
- `mergeItems` (+ test)
- `recipeIds` (+ test)
- `foodCategories`
- `nutritionFormat` (тепер бере `toLocalISODate` напряму з `@sergeant/shared`)

Щоб не ламати існуючі імпорти `../lib/mealTypes.js` / `../lib/…` у
`apps/web/src/modules/nutrition`, замість старих файлів лишаються
тонкі реекспорт-шими (той самий патерн що у `nutrition/lib/macros.ts`).

Решта шарів (storage, cloud-backup, IndexedDB recipe-book, stats /
day-summary — наразі зв'язані з `createModuleStorage` / `localStorage`)
залишаються у `apps/web` і приходять наступними PR-ами після того
як витягнемо platform-IO за `KVStore`-абстракцію.

Перевірено: `pnpm turbo run test typecheck lint` для
`@sergeant/{shared,nutrition-domain,web}` — 9/9 green,
617 web-тестів + 20 нових тестів пакета.
