+++ apps/web/package.json
@@ -25,6 +25,7 @@
     "@sergeant/mobile-shell": "workspace:*",
     "@sergeant/finyk-domain": "workspace:*",
     "@sergeant/fizruk-domain": "workspace:*",
+    "@sergeant/nutrition-domain": "workspace:*",
     "@sergeant/routine-domain": "workspace:*",
     "@sergeant/insights": "workspace:*",
     "@sergeant/mobile-shell": "workspace:*",
