+++ packages/nutrition-domain/package.json
@@ -1,0 +1,32 @@
+{
+  "name": "@sergeant/nutrition-domain",
+  "version": "0.1.0",
+  "private": true,
+  "type": "module",
+  "main": "./src/index.ts",
+  "types": "./src/index.ts",
+  "exports": {
+    ".": "./src/index.ts",
+    "./food-categories": "./src/foodCategories.ts",
+    "./meal-types": "./src/mealTypes.ts",
+    "./merge-items": "./src/mergeItems.ts",
+    "./nutrition-format": "./src/nutritionFormat.ts",
+    "./pantry-text-parser": "./src/pantryTextParser.ts",
+    "./recipe-ids": "./src/recipeIds.ts"
+  },
+  "scripts": {
+    "typecheck": "tsc -p tsconfig.json --noEmit",
+    "lint": "eslint \"src/**/*.{ts,tsx}\"",
+    "test": "vitest run",
+    "test:watch": "vitest"
+  },
+  "dependencies": {
+    "@sergeant/shared": "workspace:*"
+  },
+  "devDependencies": {
+    "@sergeant/config": "workspace:*",
+    "@types/node": "^25.6.0",
+    "typescript": "^6.0.3",
+    "vitest": "^3.0.0"
+  }
+}
