+++ packages/nutrition-domain/src/pantryTextParser.test.ts
